seperate web3 logic
